renames OpenEyePreparation.py to OpenEyeModeling.py
